+++ build.gradle
@@ -25,8 +25,12 @@ ext {
 
 repositories {
     mavenCentral()
-    mavenLocal()
-}
+    maven {
+        url 'https://esri.jfrog.io/artifactory/arcgis'
+    }
+    maven {
+        url 'https://olympus.esri.com/artifactory/arcgisruntime-repo'
+    }}
 
 configurations {
     natives
