+++ src/main/resources/main.fxml
@@ -1,0 +1,75 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ Copyright 2021 Esri.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License"); you may not
+  ~ use this file except in compliance with the License. You may obtain a copy of
+  ~ the License at
+  ~
+  ~ http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+  ~ WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+  ~ License for the specific language governing permissions and limitations under
+  ~ the License.
+  -->
+<?import javafx.scene.layout.StackPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.chart.LineChart?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.chart.NumberAxis?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.Label?>
+<!--<StackPane fx:controller="com.mbcoder.iot.weatherstation.Controller" xmlns:fx="http://javafx.com/fxml"-->
+<!--        stylesheets="@style.css" VBox.vgrow="ALWAYS">-->
+
+    <BorderPane fx:controller="com.mbcoder.iot.weatherstation.Controller" xmlns:fx="http://javafx.com/fxml"
+                stylesheets="@style.css" VBox.vgrow="ALWAYS" maxHeight="Infinity">
+        <top>
+            <HBox fx:id="hBox" styleClass="panel-region" alignment="CENTER" spacing="5">
+                <CheckBox fx:id="chkSimulated" text="Simulated" selected="true"/>
+                <Button fx:id="btnConnectSensor" onAction="#handleSensorConnectButton" text="Connect sensor"/>
+                <Button fx:id="btnLogWeather" onAction="#startWeatherLogging" text="Log weather"/>
+            </HBox>
+        </top>
+
+    <left>
+        <Label fx:id="labelTemp" text="..." styleClass="panel-upper"/>
+
+    </left>
+    <right>
+        <Label fx:id="labelPressure" text="..." styleClass="panel-upper"/>
+
+    </right>
+
+        <bottom>
+            <!--Controls box-->
+            <VBox fx:id="vBox" StackPane.alignment="BOTTOM_CENTER" maxWidth="Infinity" maxHeight="Infinity" styleClass="panel-region">
+                <LineChart fx:id="tempChart" title="Profile" animated="false">
+                    <xAxis>
+                        <NumberAxis fx:id="xAxisTemp" side="BOTTOM" label="Observation" tickUnit="0"/>
+                    </xAxis>
+                    <yAxis>
+                        <NumberAxis fx:id="yAxisTemp" side="LEFT" label="Temp (C)" tickUnit="0"/>
+                    </yAxis>
+                </LineChart>
+                <LineChart fx:id="pressureChart" title="Profile" animated="false">
+                    <xAxis>
+                        <NumberAxis fx:id="xAxisPressure" side="BOTTOM" label="Observation" tickUnit="0"/>
+                    </xAxis>
+                    <yAxis>
+                        <NumberAxis fx:id="yAxisPressure" side="LEFT" label="Pressure" tickUnit="0"/>
+                    </yAxis>
+                </LineChart>
+            </VBox>
+        </bottom>
+    </BorderPane>
+
+<!--</StackPane>-->
+
+
+
