+++ src/main/resources/style.css
@@ -1,0 +1,22 @@
+VBox {
+    -fx-background-color: rgba(0, 0, 0, 0.6);
+    -fx-padding: 20;
+    -fx-spacing: 5;
+    -fx-alignment: top-left;
+    -fx-text-fill: white;
+    -fx-max-height: 300;
+
+}
+
+.panel-region .label {
+  -fx-text-fill: white;
+  -fx-font-family: "Arial";
+  -fx-font-size: 12;
+}
+
+.panel-upper .label {
+  -fx-text-fill: white;
+  -fx-font-family: "Arial";
+  -fx-font-size: 20;
+}
+
